refactor(shared): extract constants and types into dedicated modules
constants.ts holds all domain limits and UI-status thresholds, types.ts the
shared interfaces. index.ts re-exports both, so consumers don't change.
